requirements_pdf.txt
pdfplumber==0.10.3
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
icalendar==5.0.11
pytz==2023.3
